Renamed Functions project for consistency with the WebHost version
