+++ src/Microsoft.Bot.Builder.Runtime.Functions/Microsoft.Bot.Builder.Runtime.FunctionHost.csproj
