+++ src/Microsoft.Bot.Builder.Runtime.Functions/Startup.cs
@@ -2,13 +2,12 @@
 // Licensed under the MIT License.
 
 using System;
-using Microsoft.AspNetCore.Hosting;
 using Microsoft.Azure.Functions.Extensions.DependencyInjection;
 using Microsoft.Bot.Builder.Runtime.Extensions;
 
-[assembly: FunctionsStartup(typeof(Microsoft.Bot.Builder.Runtime.Functions.Startup))]
+[assembly: FunctionsStartup(typeof(Microsoft.Bot.Builder.Runtime.FunctionHost.Startup))]
 
-namespace Microsoft.Bot.Builder.Runtime.Functions
+namespace Microsoft.Bot.Builder.Runtime.FunctionHost
 {
     public class Startup : FunctionsStartup
     {
@@ -19,7 +18,7 @@ namespace Microsoft.Bot.Builder.Runtime.Functions
 
         public override void ConfigureAppConfiguration(IFunctionsConfigurationBuilder configurationBuilder)
         {
-            var isDevelopment = string.Equals(configurationBuilder.GetContext().EnvironmentName, EnvironmentName.Development, StringComparison.OrdinalIgnoreCase);
+            var isDevelopment = string.Equals(configurationBuilder.GetContext().EnvironmentName, Microsoft.Extensions.Hosting.Environments.Development, StringComparison.OrdinalIgnoreCase);
             var applicationRoot = configurationBuilder.GetContext().ApplicationRootPath;
 
             configurationBuilder.ConfigurationBuilder.AddBotRuntimeConfiguration(applicationRoot, isDevelopment);
